requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas~=1.3.5
aicrowd-cli
runipy
IPython
seaborn
scipy~=1.8.1
matplotlib
numpy==1.21
polar
kaggle
tqdm~=4.64.0
sklearn~=0.0
scikit-learn~=1.1.2
setuptools~=65.3.0